Revert "1.1"
This reverts commit 4f193c60d7c6d62b77257bc10a921298ca230eac.

+++ src/CrewList.cpp
@@ -119,7 +119,7 @@ void CrewList::loadData()
 		wxStringTokenizer tkz(line, _T("\t"),wxTOKEN_RET_EMPTY);
 		int c;
 		int colCount = gridCrew->GetNumberCols();
-		if(tkz.CountTokens() != (unsigned int) colCount)
+		if(tkz.CountTokens() != colCount)
 			{ c = 1; newCol = true; }
 		else
 			{ c = 0; newCol = false; }
